test: device,inode: compact, stricter

+++ +stdlib/+java/device.m
@@ -1,6 +1,6 @@
 function i = device(p)
 
 opt = javaMethod("values", "java.nio.file.LinkOption");
-i = javaMethod("getAttribute", "java.nio.file.Files", javaPathObject(p), "unix:dev", opt);
+i = uint64(javaMethod("getAttribute", "java.nio.file.Files", javaPathObject(p), "unix:dev", opt));
 
 end
